revert changes from fa36d8bbcfe00a95a5dfd52 that broke backward compatibility

+++ sys/net/netmap_legacy.h
@@ -152,7 +152,7 @@ struct nmreq {
 	char		nr_name[IFNAMSIZ];
 	uint32_t	nr_version;	/* API version */
 	uint32_t	nr_offset;	/* nifp offset in the shared region */
-	size_t	        nr_memsize;	/* size of the shared region */
+	uint32_t	nr_memsize;	/* size of the shared region */
 	uint32_t	nr_tx_slots;	/* slots in tx rings */
 	uint32_t	nr_rx_slots;	/* slots in rx rings */
 	uint16_t	nr_tx_rings;	/* number of tx rings */
